Moves the build logic to Bit Engine
