+++ src/E5R.Tools.Bit.Engine/E5R.Tools.Bit.Engine.csproj
@@ -5,6 +5,11 @@
     <AssemblyName>E5R.Tools.Bit.Engine</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="2.3.2" />
+    <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.0.0" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\E5R.Sdk.Bit\E5R.Sdk.Bit.csproj" />
   </ItemGroup>
 </Project>
