+++ src/E5R.Tools.Bit/E5R.Tools.Bit.csproj
@@ -6,9 +6,7 @@
     <AssemblyName>bit</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="2.3.2" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.0.0" />
-    <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\E5R.Sdk.Bit\E5R.Sdk.Bit.csproj" />
